requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
numpy>=1.21.0
tqdm>=4.64.0
pillow>=8.0.0
